requirements.txt
pdfplumber
pandas
python-dotenv
google-generativeai
groq
pytest
